requirements.txt
beautifulsoup4==4.12.0
chardet==5.2.0
diskcache==5.6.3
httpx==0.27.0
lxml==4.6.2
numpy>=1.23.3,<1.24
openai==1.23.2
protobuf==3.20.0
python-dotenv==1.0.0
Requests==2.31.0
together==1.1.2
tqdm==4.66.2
